Moved StringTemplate to the `util` package. Finished the ResumeExporter's file-generating capabilities.
