+++ src/main/java/to/us/resume_builder/export/ILaTeXConvertable.java
@@ -1,7 +1,16 @@
 package to.us.resume_builder.export;
 
-import to.us.resume_builder.export.template.ResumeTemplate;
-
+/**
+ * Implementing this interface indicates that the class is LaTeX-serializable
+ * through the use of a {@link ResumeTemplate}.
+ */
 public interface ILaTeXConvertable {
+    /**
+     * Get the result of serializing this object using the specified template.
+     *
+     * @param template The template to format this object with.
+     *
+     * @return A String representing the object in the LaTeX template.
+     */
     String formatLaTeXString(ResumeTemplate template);
 }
