--- src/main/java/to/us/resume_builder/export/ResumeExportException.java
@@ -1,7 +1,0 @@
-package to.us.resume_builder.export;
-
-public class ResumeExportException extends Exception {
-    public ResumeExportException(String message) {
-        super(message);
-    }
-}
