+++ src/main/java/to/us/resume_builder/export/ResumeExporter.java
@@ -1,27 +1,35 @@
 package to.us.resume_builder.export;
 
-import to.us.resume_builder.export.template.ResumeTemplate;
-import to.us.resume_builder.export.template.StringTemplate;
 import to.us.resume_builder.resume_components.Resume;
 import to.us.resume_builder.resume_components.category.Category;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
 import java.util.StringJoiner;
 
+/**
+ * This class handles the exporting of a resume file to a PDF.
+ *
+ * @author Matthew McCaskill
+ * @see Resume
+ * @see ResumeTemplate
+ */
 public class ResumeExporter {
     private Resume resume;
-
+    
     public ResumeExporter(Resume resume) {
         this.resume = resume;
     }
 
     public boolean export(String fileName) throws IOException {
+        return export(fileName, ResumeTemplate.DEFAULT);
+    }
+
+    public boolean export(String fileName, ResumeTemplate template) throws IOException {
         String latexCode;
-        latexCode = getLaTeXString(ResumeTemplate.DEFAULT);
+        latexCode = getLaTeXString(template);
 
         // TODO: remove; debug
         Files.writeString(Path.of("export.txt"), latexCode);
