+++ src/main/java/to/us/resume_builder/resume_components/category/BulletCategory.java
@@ -1,8 +1,6 @@
 package to.us.resume_builder.resume_components.category;
 
-import to.us.resume_builder.export.ResumeExportException;
-import to.us.resume_builder.export.template.ResumeTemplate;
-import to.us.resume_builder.export.template.StringTemplate;
+import to.us.resume_builder.export.ResumeTemplate;
 import to.us.resume_builder.resume_components.Field;
 
 import java.util.LinkedList;
