+++ src/main/java/to/us/resume_builder/resume_components/category/ExperienceCategory.java
@@ -1,7 +1,6 @@
 package to.us.resume_builder.resume_components.category;
 
-import to.us.resume_builder.export.template.ResumeTemplate;
-import to.us.resume_builder.export.template.StringTemplate;
+import to.us.resume_builder.export.ResumeTemplate;
 import to.us.resume_builder.resume_components.Experience;
 
 import java.util.LinkedList;
