+++ src/main/java/to/us/resume_builder/resume_components/category/HeaderCategory.java
@@ -1,8 +1,7 @@
 package to.us.resume_builder.resume_components.category;
 
 
-import to.us.resume_builder.export.template.ResumeTemplate;
-import to.us.resume_builder.export.template.StringTemplate;
+import to.us.resume_builder.export.ResumeTemplate;
 
 public class HeaderCategory extends Category {
